Remove CartoDataFrame class. Move managers to utils
