+++ cartoframes/utils/managers/context_manager.py
@@ -1,20 +1,21 @@
 import time
-import pandas as pd
 
+from pandas import read_csv
 from warnings import warn
 
 from carto.auth import APIKeyAuthClient
 from carto.exceptions import CartoException, CartoRateLimitException
 from carto.sql import SQLClient, BatchSQLClient, CopySQLClient
 
+from ..logger import log
+from ..geom_utils import encode_geometry_ewkb
+from ..utils import is_sql_query, check_credentials, encode_row, map_geom_type, PG_NULL
+from ..columns import Column, get_dataframe_columns_info, obtain_converters, \
+                      date_columns_names, normalize_name
+
 from ... import __version__
-from ...utils.logger import log
 from ...io.dataset_info import DatasetInfo
 from ...auth.defaults import get_default_credentials
-from ...utils.geom_utils import encode_geometry_ewkb
-from ...utils.utils import is_sql_query, check_credentials, encode_row, map_geom_type, PG_NULL
-from ...utils.columns import Column, get_dataframe_columns_info, obtain_converters, \
-                             date_columns_names, normalize_name
 
 
 DEFAULT_RETRY_TIMES = 3
@@ -265,7 +266,7 @@ class ContextManager:
         converters = obtain_converters(columns)
         parse_dates = date_columns_names(columns)
 
-        df = pd.read_csv(
+        df = read_csv(
             raw_result,
             converters=converters,
             parse_dates=parse_dates)
