--- docs/includes/cartodataframe.rst
@@ -1,9 +1,0 @@
-CartoDataFrame
---------------
-
-.. automodule:: cartoframes.core.cartodataframe
-    :members:
-    :exclude-members: astype, merge, dissolve, explode
-    :member-order: bysource
-    :undoc-members:
-    :show-inheritance:
